implemented confirmation functionality

+++ frontend/src/Hooks/ArticleHooks/UseDeleteMany.js
@@ -1,24 +1,40 @@
 import { useState } from "react";
 import { useGlobalErrorContext } from "../../Contexts/GlobalErrorContext";
 import articleApi from "../../Utils/ArticleApiFunctions";
+import { useConfirmationContext } from "../../Contexts/UseConfirmationContext";
 
 export const useDeleteMany = () => {
   const [, setGlobalError] = useGlobalErrorContext();
   const [isLoading, setIsLoading] = useState(null);
+  const { confirmation, setConfirmation } = useConfirmationContext();
 
   const deleteMany = async (ids) => {
-    setIsLoading(true);
+    confirmation.ref.current.show();
+
+    setConfirmation({
+      ...confirmation,
+      type: "deleteMany",
+      info: {
+        size: ids.length,
+      },
+    });
     setGlobalError(null);
+    const isConfirmed = await confirmation.confirmed;
+    if (isConfirmed) {
+      setIsLoading(true);
 
-    const response = await articleApi.deleteMany(ids);
-    const json = await response.json();
+      const response = await articleApi.deleteMany(ids);
+      const json = await response.json();
 
-    if (!response.ok) {
-      setGlobalError(json.message);
+      if (!response.ok) {
+        setGlobalError(json.message);
+      }
+      setIsLoading(false);
+
+      return response;
+    } else {
+      return;
     }
-
-    setIsLoading(false);
-    return response;
   };
 
   return { deleteMany, isLoading };
